feat: add request and response models for User
Created request and response structs to handle User data in API operations.

+++ src/configuration/rest_err/rest_err.go
@@ -7,7 +7,7 @@ import (
 type RestErr struct {
 	Message string  `json:"message"`
 	Err     string  `json:"error"`
-	Code    int64   `json:"code"`
+	Code    int     `json:"code"`
 	Causes  []Cause `json:"causes"`
 }
 
@@ -20,7 +20,7 @@ func (r *RestErr) Error() string {
 	return r.Message
 }
 
-func NewRestErr(message, err string, code int64, causes []Cause) *RestErr {
+func NewRestErr(message, err string, code int, causes []Cause) *RestErr {
 	return &RestErr{
 		Message: message,
 		Err:     err,
